app/requirements.txt
fastapi
uvicorn[standard]
psycopg2
passlib
python-jose
behave
pika
sqlalchemy